Changed the command for automatically creating event tree diagrams
